requirements.txt
mkdocs
mkdocs-material
plotly
PyYAML